# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateDimensionRequest:
    properties:
      columnName:
        type: string
      dataType:
        type: string
      description:
        type: string
      isHidden:
        type: boolean
      name:
        type: string
    type: object
  handlers.CreateMetricRequest:
    properties:
      description:
        type: string
      format:
        type: string
      formula:
        type: string
      name:
        type: string
    type: object
  handlers.CreateSemanticModelRequest:
    properties:
      dataSourceId:
        type: string
      description:
        type: string
      dimensions:
        items:
          $ref: '#/definitions/handlers.CreateDimensionRequest'
        type: array
      metrics:
        items:
          $ref: '#/definitions/handlers.CreateMetricRequest'
        type: array
      name:
        type: string
      tableName:
        type: string
    type: object
...
  handlers.SemanticQueryRequest:
    properties:
      dimensions:
        items:
          type: string
        type: array
      filters:
        additionalProperties: true
        type: object
      limit:
        type: integer
      metrics:
        items:
          type: string
        type: array
      modelId:
        type: string
    type: object
  handlers.SemanticQueryResponse:
    properties:
      args:
        items: {}
        type: array
      dimensions:
        items:
          type: string
        type: array
      metrics:
        items:
          type: string
        type: array
      rowCount:
        type: integer
      sql:
        type: string
    type: object
...
  models.SemanticDimension:
    properties:
      columnName:
        type: string
      createdAt:
        type: string
      dataType:
        description: string, number, date, boolean
        type: string
      description:
        type: string
      id:
        type: string
      isHidden:
        type: boolean
      modelId:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.SemanticMetric:
    properties:
      createdAt:
        type: string
      description:
        type: string
      format:
        description: currency, percentage, number, etc.
        type: string
      formula:
        description: e.g., "SUM(revenue)", "COUNT(*)", "AVG(price)"
        type: string
      id:
        type: string
      modelId:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.SemanticModel:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      dataSourceId:
        type: string
      description:
        type: string
      dimensions:
        items:
          $ref: '#/definitions/models.SemanticDimension'
        type: array
      id:
        type: string
      metrics:
        items:
          $ref: '#/definitions/models.SemanticMetric'
        type: array
      name:
        type: string
      tableName:
        type: string
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@insightengine.ai
    name: API Support
  description: Enterprise BI Platform API with Power BI/Tableau parity.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: InsightEngine API
  version: "1.0"
paths:
...
  /api/v1/semantic/models:
    get:
      consumes:
      - application/json
      description: Get all semantic models for the user's workspace
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SemanticModel'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List semantic models
      tags:
      - semantic-layer
    post:
      consumes:
      - application/json
      description: Create a new semantic model with dimensions and metrics
      parameters:
      - description: Model data
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSemanticModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SemanticModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create semantic model
      tags:
      - semantic-layer
  /api/v1/semantic/models/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a semantic model
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete semantic model
      tags:
      - semantic-layer
    get:
      consumes:
      - application/json
      description: Get a semantic model by ID
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SemanticModel'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get semantic model details
      tags:
      - semantic-layer
    put:
      consumes:
      - application/json
      description: Update an existing semantic model
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      - description: Model data
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSemanticModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SemanticModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update semantic model
      tags:
      - semantic-layer
  /api/v1/semantic/query:
    post:
      consumes:
      - application/json
      description: Execute a query using business terms (dimensions and metrics)
      parameters:
      - description: Query data
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/handlers.SemanticQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SemanticQueryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Execute semantic query
      tags:
      - semantic-layer
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"